Fix strings reloading after a restart
When replaying blocks from genesis, the first `commit` is called
without a `checkout` before.
The implementation was assuming that `commit` was always following a
`checkout`

Also fix the big strings file, on reload it was copying its already
existing content to the end of file. The size of the file was doubling
on each restart.